+++ Vostok.Metrics.System/Host/HostMetricsCollectorExtensions.cs
@@ -1,11 +1,8 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Net.NetworkInformation;
 using JetBrains.Annotations;
 using Vostok.Metrics.Models;
 using Vostok.Metrics.Primitives.Gauge;
-using Vostok.Metrics.System.Helpers;
 
 namespace Vostok.Metrics.System.Host
 {
@@ -28,54 +25,7 @@ namespace Vostok.Metrics.System.Host
         {
             var metrics = collector.Collect();
 
-            foreach (var property in typeof(HostMetrics).GetProperties())
-            {
-                if (property.PropertyType.GetInterface(nameof(IDictionary)) == null)
-                {
-                    yield return new MetricDataPoint(
-                        Convert.ToDouble(property.GetValue(metrics)),
-                        (WellKnownTagKeys.Name, property.Name));
-                }
-            }
-
-            foreach (var tcpState in metrics.TcpStates.OrEmptyIfNull())
-            {
-                yield return new MetricDataPoint(
-                    Convert.ToDouble(tcpState.Value),
-                    (WellKnownTagKeys.Name, "TcpConnectionCountPerState"),
-                    (nameof(TcpState), tcpState.Key.ToString())
-                );
-            }
-
-            foreach (var diskSpaceInfo in metrics.DisksSpaceInfo.OrEmptyIfNull())
-            {
-                foreach (var property in typeof(DiskSpaceInfo).GetProperties())
-                {
-                    if (!property.Name.Equals(nameof(DiskSpaceInfo.DiskName)) && !property.Name.Equals(nameof(DiskSpaceInfo.RootDirectory)))
-                    {
-                        yield return new MetricDataPoint(
-                            Convert.ToDouble(property.GetValue(diskSpaceInfo.Value)),
-                            (WellKnownTagKeys.Name, $"Disk{property.Name}"),
-                            (nameof(DiskSpaceInfo.DiskName), diskSpaceInfo.Value.DiskName)
-                        );
-                    }
-                }
-            }
-
-            foreach (var diskUsageInfo in metrics.DisksUsageInfo.OrEmptyIfNull())
-            {
-                foreach (var property in typeof(DiskUsageInfo).GetProperties())
-                {
-                    if (!property.Name.Equals(nameof(DiskUsageInfo.DiskName)))
-                    {
-                        yield return new MetricDataPoint(
-                            Convert.ToDouble(property.GetValue(diskUsageInfo.Value)),
-                            (WellKnownTagKeys.Name, $"Disk{property.Name}"),
-                            (nameof(DiskSpaceInfo.DiskName), diskUsageInfo.Value.DiskName)
-                        );
-                    }
-                }
-            }
+            return metrics.ToDataPoints();
         }
     }
 }
