+++ Vostok.Metrics.System/Process/CurrentProcessMetricsCollectorExtensions.cs
@@ -20,13 +20,7 @@ namespace Vostok.Metrics.System.Process
         {
             var metrics = collector.Collect();
 
-            foreach (var property in typeof(CurrentProcessMetrics).GetProperties())
-            {
-                var value = property.GetValue(metrics);
-
-                if (value != null)
-                    yield return new MetricDataPoint(Convert.ToDouble(value), (WellKnownTagKeys.Name, property.Name));
-            }
+            return metrics.ToDataPoints();
         }
     }
 }
