+++ Vostok.Metrics.System/Process/CurrentProcessMonitorExtensions.cs
@@ -10,31 +10,7 @@ namespace Vostok.Metrics.System.Process
     {
         private static Action<ILog, TimeSpan, CurrentProcessMetrics> loggingRule = (log, period, metrics) =>
         {
-            log.Info(
-                "CPU = {CpuUsagePercent:0.00}% ({CpuUsageCores:0.00} cores). " +
-                "Memory = {MemoryResidentUsage} ({GcHeapSize} managed). " +
-                "ThreadPool: {ThreadPoolTotalCount} total, {ThreadPoolBusyWorkers}/{ThreadPoolMinWorkers} workers, " +
-                "{ThreadPoolBusyIo}/{ThreadPoolMinIo} IO, {ThreadPoolQueueLength} queue. " +
-                "Handles = {HandlesCount}. Timers = {TimersCount}. " +
-                "Alloc/s = {AllocatedPerSecond:0.00}. " +
-                "Contentions/s = {ContentionsPerSecond:0.00}. " +
-                "Exceptions/s = {ExceptionsPerSecond:0.00}.",
-                metrics.CpuUtilizedFraction * 100,
-                metrics.CpuUtilizedCores,
-                SizeFormatter.Format(metrics.MemoryResident),
-                SizeFormatter.Format(metrics.GcHeapSize),
-                metrics.ThreadPoolTotalCount,
-                metrics.ThreadPoolBusyWorkers,
-                metrics.ThreadPoolMinWorkers,
-                metrics.ThreadPoolBusyIo,
-                metrics.ThreadPoolMinIo,
-                metrics.ThreadPoolQueueLength,
-                metrics.HandlesCount,
-                metrics.ActiveTimersCount,
-                SizeFormatter.Format((long) (metrics.GcAllocatedBytes / period.TotalSeconds)),
-                metrics.LockContentionCount / period.TotalSeconds,
-                metrics.ExceptionsCount / period.TotalSeconds
-            );
+            metrics.LogMetrics(log, period);
         };
 
         /// <summary>
